removed model too specific, added simple varparsing for physics options
